programs/user_profile_picture_uploader/3-PnS-haiku4.5/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
python-magic-bin==0.4.14
